e2e-demos/external-lookup/pyproject.toml
[project]
name = "external-lookup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.2.15",
]
